Update README.md
the following actions have to be  performed in order to develop a machine learning prototype model for classifying flower plants:

Gather and label data: You'll need a dataset with pictures of flower plants and the names that go with them  flowers or plants. You have the option of gathering this data yourself or using an existing dataset. So that the model may learn to link particular characteristics of each flower plant with its label, the data should be labelled.

Data preprocessing is necessary in order to feed the data into a machine learning model. The photos must be resized, formatted in a common manner, and have their pixel values normalized.